create separate workflow for sdown
